cleaning (remove tablayout move to only 1 activity)
