+++ .idea/misc.xml
@@ -20,11 +20,13 @@
         <entry key="..\:/Code/Android/MyCards/app/src/main/res/layout/label_text_field.xml" value="0.4" />
         <entry key="..\:/Code/Android/MyCards/app/src/main/res/layout/no_label_date_field.xml" value="0.4" />
         <entry key="..\:/Code/Android/MyCards/app/src/main/res/layout/no_label_text_field.xml" value="0.4" />
-        <entry key="..\:/Code/My-Cards/app/src/main/res/layout/activity_add_membership.xml" value="0.2" />
+        <entry key="..\:/Code/My-Cards/app/src/main/res/layout/activity_add_membership.xml" value="0.4" />
+        <entry key="..\:/Code/My-Cards/app/src/main/res/layout/activity_main.xml" value="0.20863616930312098" />
         <entry key="..\:/Code/My-Cards/app/src/main/res/layout/fragment_add_card_tab.xml" value="0.3252032520325203" />
         <entry key="..\:/Code/My-Cards/app/src/main/res/layout/fragment_add_coupon_tab.xml" value="0.13100634632819583" />
-        <entry key="..\:/Code/My-Cards/app/src/main/res/layout/fragment_add_sub_tab.xml" value="0.13780598368087035" />
+        <entry key="..\:/Code/My-Cards/app/src/main/res/layout/fragment_add_sub_tab.xml" value="0.264" />
         <entry key="..\:/Code/My-Cards/app/src/main/res/layout/item_label_date_input.xml" value="0.264" />
+        <entry key="..\:/Code/My-Cards/app/src/main/res/layout/item_label_text_input.xml" value="0.20863616930312098" />
         <entry key="..\:/Code/My-Cards/app/src/main/res/layout/no_label_date_field.xml" value="0.72" />
         <entry key="..\:/Code/My-Cards/app/src/main/res/layout/no_label_text_field.xml" value="0.6" />
       </map>
