+++ app/src/main/java/com/example/mycards/MainActivity.java
@@ -1,19 +1,11 @@
 package com.example.mycards;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-
-import com.example.mycards.controller.CustomStringRecyclerAdapter;
-import com.example.mycards.model.Pair;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
