+++ app/src/main/java/com/example/mycards/controller/util/ItemClickListener.java
@@ -1,4 +1,4 @@
-package com.example.mycards.controller;
+package com.example.mycards.controller.util;
 
 import android.view.View;
