--- app/src/main/java/com/example/mycards/fragments/addCouponTab.java
@@ -1,21 +1,0 @@
-package com.example.mycards.fragments;
-
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.example.mycards.R;
-
-
-public class addCouponTab extends Fragment {
-
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_add_coupon_tab, container,false);
-    }
-}
