--- app/src/main/res/layout/fragment_add_sub_tab.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_marginTop="?attr/actionBarSize"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".fragments.addSubTab">
-
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:text="H3"
-        android:textSize="34sp" />
-</FrameLayout>
